src/test/java/Aspose/Slides/Cloud/Sdk/Extra/PipelineTest.java
/*
 * --------------------------------------------------------------------------------------------------------------------
 * <copyright company="Aspose">
 *   Copyright (c) 2018 Aspose.Slides for Cloud
 * </copyright>
 * <summary>
 *   Permission is hereby granted, free of charge, to any person obtaining a copy
 *  of this software and associated documentation files (the "Software"), to deal
 *  in the Software without restriction, including without limitation the rights
 *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 *  copies of the Software, and to permit persons to whom the Software is
 *  furnished to do so, subject to the following conditions:
 * 
 *  The above copyright notice and this permission notice shall be included in all
 *  copies or substantial portions of the Software.
 * 
 *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 *  SOFTWARE.
 * </summary>
 * --------------------------------------------------------------------------------------------------------------------
 */

package Aspose.Slides.Cloud.Sdk.Extra;

import Aspose.Slides.Cloud.Sdk.Api.DocumentApi;
import Aspose.Slides.Cloud.Sdk.ApiException;
import org.junit.Test;

import Aspose.Slides.Cloud.Sdk.ApiTest;
import Aspose.Slides.Cloud.Sdk.FileInfo;
import Aspose.Slides.Cloud.Sdk.Model.*;
import Aspose.Slides.Cloud.Sdk.Model.Request.PostSlidesPipelineRequest;
import java.io.File;
import java.io.IOException;
import java.nio.file.Files;
import java.nio.file.Paths;
import java.util.ArrayList;
import java.util.List;
import static org.junit.Assert.assertTrue;

/**
 * API tests for PlaceholdersApi
 */
public class PipelineTest extends ApiTest {

    private final DocumentApi api = new DocumentApi(getConfiguration());

    
    /**
     * Read slide placeholder info.
     *
     * 
     *
     * @throws ApiException
     *          if the Api call fails
     */
    @Test
    public void placeholderTest() throws ApiException, IOException {
        RequestInputFile file1 = new RequestInputFile();
        file1.setType(InputFileType.REQUEST);
        file1.setIndex(0);

        RequestInputFile file2 = new RequestInputFile();
        file2.setType(InputFileType.REQUEST);
        file2.setIndex(1);

        Input input = new Input();
        input.setTemplateData(file1);
        input.setTemplate(file2);

        OutputFile output = new OutputFile();
        output.setType(OutputFileType.RESPONSE);

        Save task = new Save();
        task.setFormat(ExportFormat2.PPTX);
        task.setOutput(output);
        task.setType(TaskType.SAVE);
        List<Task> tasks = new ArrayList<Task>();
        tasks.add(task);

        Pipeline pipeline = new Pipeline();
        pipeline.setInput(input);
        pipeline.setTasks(tasks);

        FileInfo fileInfo1 = new FileInfo();
        fileInfo1.setMimeType("text/xml");
        fileInfo1.setData(Files.readAllBytes(Paths.get("TestData/TemplatingCVDataWithBase64.xml")));

        FileInfo fileInfo2 = new FileInfo();
        fileInfo2.setName("TemplateCV.pptx");
        fileInfo2.setData(Files.readAllBytes(Paths.get("TestData/TemplateCV.pptx")));
        
        List<FileInfo> files = new ArrayList<FileInfo>();
        files.add(fileInfo1);
        files.add(fileInfo2);
        
        PostSlidesPipelineRequest request = new PostSlidesPipelineRequest();
        request.setPipeline(pipeline);
        request.setFiles(files);
                
        File response = api.postSlidesPipeline(request);
        assertTrue(response.isFile());
    }
}
